refactor: 헤더전송 메소드와 바디 전송 메소드 분리
